# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation Moon Group",
        "description": "API Documentation for Moon Group",
        "version": "1.0.0"
    },
    "paths": {
...
        "/moon-group-backend/public/api/contact": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Get all contact messages",
                "description": "Get all contact messages",
                "operationId": "086392b5b7cb325cfb996dcff6ed0bb3",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Send a message",
                "description": "Send a message",
                "operationId": "920153db77b5bd341711017188b2ca8a",
                "requestBody": {
                    "description": "Pass contact details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Message sent successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "The pagination must be an integer."
                    }
                },
                "type": "object"
            },
            "Unauthenticated": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "StoreContactRequest": {
                "required": [
                    "name",
                    "subject",
                    "email"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Subject"
                    },
                    "email": {
                        "type": "string",
                        "example": "mail@gmail.com"
                    },
                    "message": {
                        "type": "string",
                        "example": "Message"
                    }
                },
                "type": "object"
            },
...
            "Contact": {
                "required": [
                    "name",
                    "subject",
                    "email",
                    "message"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Subject"
                    },
                    "email": {
                        "type": "string",
                        "example": "mail@gmail.com"
                    },
                    "message": {
                        "type": "string",
                        "example": "Message"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}